MoodEvents location on the Map
I am done connecting my Map to the Moods location. So basically it takes every moods and puts it in the Map. For now I am using the main Hashmap but can change later based on filtering. Still awating on profile list so that I can make Maps based on user location.

Fixed the bug with the location fetching. Idk why but android studio thinks I am in the US (Maybe because the Emulators servers are there, I can see the google headquarter in the google maps in my location lol)